+++ mainService/src/main/java/com/threestar/selectstar/controller/MeetingController.java
@@ -43,7 +43,7 @@ public class MeetingController {
     }
     // 단건 조회
     @GetMapping("/{id}")
-    public ResponseEntity<FindMeetingOneResponse> meetingDetail(@PathVariable("id") int id,@AuthenticationPrincipal CustomUserDetails userDetails){
+    public ResponseEntity<FindMeetingOneResponse> meetingDetail(@PathVariable("id") Long id,@AuthenticationPrincipal CustomUserDetails userDetails){
         FindMeetingOneResponse meetingOne = meetingService.findMeetingOne(id);
 
         meetingOne.setImg(userService.getUserProfile(meetingOne.getUserId()).getProfilePhoto());
@@ -93,7 +93,7 @@ public class MeetingController {
     }
     
     @DeleteMapping("/{id}")
-    public Map<String,String> meetingRemove(@PathVariable("id") int id){
+    public Map<String,String> meetingRemove(@PathVariable("id") Long id){
         Map<String, String> succesMap = new HashMap<>();
         succesMap.put("result",meetingService.removeMeeting(id));
         return succesMap;
